requirements/readthedocs.txt
mmcv-full
munkres
poseval@git+https://github.com/svenkreiss/poseval.git
regex
scipy
titlecase
torch
torchvision
xtcocotools
